etl_pipeline/requirements.txt
# Data handling
pandas==1.3.5
openpyxl==3.1.3
tqdm==4.66.4
regex==2024.4.16

# Embedding & NLP
sentence-transformers==2.2.2
transformers==4.30.2
scikit-learn==1.0.2
nltk==3.8.1
torch==1.13.1
sentencepiece==0.2.0

# Database
psycopg2==2.9.9

# Web scraping
beautifulsoup4==4.12.3
lxml==5.4.0
requests==2.31.0

# Google Cloud
google-cloud-storage==3.1.0
google-auth==2.40.1
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-api-core==2.24.2

# Dashboard / Visualisasi
streamlit==1.23.1
altair==5.0.1

# Misc (jaga-jaga dependensi internal)
python-dotenv==0.21.1
supabase